[fix] Permit Thought update
Thought update was not permitted because of the user_publish_once_a_day_only
validation which was checking if user has published the current day.

Now if a created_at attribute exists, we search if user has published this
specific day and if this publication has a different id. Otherwise, it means
we are editing the thought.